Remove maxUnavailable config from statefulset
This feature is alpha state, see https://kubernetes.io/docs/concepts/workloads/controllers/statefulset/#maximum-unavailable-pods. We should not configure this behaviour in our default set of values. I don't know if our DO k8s api server has this feature gate enabled, so I won't configure it for the dev environment. Also, not sure if this only makes sense with podManagementPoliyc=Parallel which we haven't configured for this statefulset.