PORTAL-1379: Mount empty dir volume at /file-uploads
For reasons unknown, vertx tries to create the directory /file-uploads to store file uploads. The creation of this directory might fail due to lack of permissions by the user. The mounting of an emptyDir will bypass the creation. Also note that no actual file-uploads take place, but vertx erroneously thinks that the form-post which passes the parameters from the oauth flow is a file-upload.